refactor: rename user data upload for consistency down org data download
